html/semantics/embedded-content/the-canvas-element/toDataURL.lowercase.unicode.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: toDataURL.lowercase.unicode</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>toDataURL.lowercase.unicode</h1>
<p class="desc">toDataURL type is ASCII-case-insensitive</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("toDataURL type is ASCII-case-insensitive");
_addTest(function(canvas, ctx) {

// Use LATIN CAPITAL LETTER I WITH DOT ABOVE (Unicode lowercase is "i")
var data = canvas.toDataURL('\u0130mage/png');
_assertMatch(data, /^data:image\/png[;,]/, "data", "/^data:image\\/png[;,]/");

var data = canvas.toDataURL('\u0130mage/jpeg');
_assertMatch(data, /^data:image\/png[;,]/, "data", "/^data:image\\/png[;,]/");


});
</script>
